+++ js/sync.js
@@ -13,24 +13,28 @@ var localcontacts = new Array();
 function isDoneSyncing(total) {
 	if(total==localcontacts.length) {
 		$("#syncall").text('Done !').removeClass('loading');
+		$('#loader').remove();
+		$('#contacts-list').fadeIn();
 	}
 }
 
 function syncPictures() {
+	
+	$('#loader').fadeIn();
+	
 	var synced = 0;
 	var error = 0;
-	$('#contacts-list').fadeIn();
 	localcontacts.forEach(function(id, index, array){
 		$.get(OC.generateUrl('apps/fbsync/setphoto/'+id)).done(function(response) {
 
 			var url = OC.generateUrl(
-				'apps/contacts/addressbook/{backend}/{addressbook}/contact/{id}/photo?maxSize=100',
-				{backend: response['backend'], addressbook: response['addressbook'], id: response['id']}
+				'apps/fbsync/getphoto/{id}/100',
+				{id: response['id']}
 			);
 			var syncstatus = synced+" contact"+(synced>1?'s':'')+" synced ("+error+" error"+(error>1?'s':'')+")";
 			var contactdivStart = '<div class="sync-contact tooltipped" title="'+response['name'];
 			var contactdivError = ': '+response['error']
-			var contactdivEnd = '"><img src="'+url+'" /></div>'
+			var contactdivEnd = '"><img src="'+url+'" height="100" width="100" /></div>'
 			$('#syncstatus').text(syncstatus);
 
 			// Error or success?
@@ -90,7 +94,6 @@ function syncPictures() {
         var url = OC.generateUrl('apps/fbsync/FBcontacts')
         $.getJSON(url).done(function (response) {
 			localcontacts=response;
-			$("#loader").remove();
 			$('#syncstatus').text(localcontacts.length-1+' contacts loaded')
         }).fail(function(){
 			$('#loading-status').text('Unexpected error...');